Simplify some shootout Makefiles
By following the naming conventions of the build system we can
simplify the `Makefile`s a bit.  This patch also avoids having to
explicitly pass the flags to `runstdtest` for `fasta`, `k-nucleotide`
and `reverse-complement`.  The only consquence of this is that it's a
requirement to run `make boot` before `make` for these benchmarks. But
that is already the case since `.depend` files are generated by with
`make boot`.

Finally, this also update the `.gitignore` with the new names of
the output/input files.

Signed-off-by: Michal Terepeta <michal.terepeta@gmail.com>